TRK-APP/the-room-knows-apis/requirements.txt
flask
flask-cors
python-dotenv
flask-jwt-extended
flask-socketio
tensorflow
numpy
pillow
eventlet